Refactor conversation management routes: replace conversations.js with userChatHist.js for improved organization and maintainability
